feat(dashboard): drive the clean timelapse from the segment mode
Make the dense-stream segment capture usable from the UI, not just curl.

Backend:
- /api/cameras now flags `segment` capability (stream + park_tuning + select_tuning).
- /api/camera/config accepts the per-camera tuning as ONE raw object and splits it
  into ParkTuning (strict — a partial object is a 400) AND SelectTuning (best-effort),
  the same split the CLI --cameras-config does, so an HTTP/dashboard-configured camera
  is segment-capable just like a CLI-seeded one. The config echo MERGES the select
  knobs back into the tuning object so a manage-form re-save round-trips them (they're
  stored separately but live in the same posted object) instead of silently dropping
  select_tuning. SelectTuning gains Serialize for the echo.

Frontend:
- Camera gains a `segment` capability; TimelapseMode/TimelapseState learn `segment`.
- The "clean timelapse" picks its method best-first: segment (the robust full-layer
  dense-stream capture) → park (the old camera-detected miner) → smooth (printer-synced).
  segment supersedes park, so a calibrated camera now records the ~100%-yield timelapse.
- The "captured" preview toggle enables during a segment run too (it publishes the same
  park filmstrip), and start/stop route through the segment slot.

Tests: select_tuning_round_trips_and_flags_segment_capability (combined tuning → segment
capable + round-trip; park-only tuning stays park-only). Dashboard typechecks; 393 pass.

+++ src/core/park.rs
@@ -126,7 +126,7 @@ pub struct SelectFrame {
 /// Knobs for [`select_park_frame`]. Distinct from [`ParkTuning`]: selection scores against
 /// the per-burst MEDIAN (not the live EMA), so its cutoffs differ. No defaults — every knob
 /// is supplied by the caller (it depends on camera/printer placement).
-#[derive(Clone, Copy, Debug, PartialEq, Deserialize)]
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize)]
 pub struct SelectTuning {
     /// Park zone = the left this-fraction of the frame.
     pub left_frac: f64,
@@ -1025,7 +1025,7 @@ mod select_tests {
                 if d <= head_hw {
                     img[row + x] = BRIGHT_HEAD;
                 } else if !sharp && d <= head_hw + 3 {
-                    let base = if (x as usize) < left_w { DARK_BACKDROP } else { BG };
+                    let base = if x < left_w { DARK_BACKDROP } else { BG };
                     img[row + x] = (BRIGHT_HEAD as f64
                         + (d - head_hw) as f64 / 3.0 * (base as f64 - BRIGHT_HEAD as f64))
                         as u8;
